requirements.txt
streamlit>=1.28.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-dotenv>=1.0.0
openai>=1.3.0
numpy>=1.21.0
pandas>=1.3.0
